+++ packages/dashboard/stories/dashboard/edge-dashboard.stories.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { IoTSiteWiseClient, IoTSiteWise } from '@aws-sdk/client-iotsitewise';
+import { IoTSiteWiseClient } from '@aws-sdk/client-iotsitewise';
 import { IoTEventsClient } from '@aws-sdk/client-iot-events';
 import { IoTTwinMakerClient } from '@aws-sdk/client-iottwinmaker';
 import { registerPlugin } from '@iot-app-kit/core';
@@ -27,7 +27,6 @@ const clientConfig = {
 const iotSiteWiseClient = new IoTSiteWiseClient(clientConfig);
 const iotEventsClient = new IoTEventsClient(clientConfig);
 const iotTwinMakerClient = new IoTTwinMakerClient(clientConfig);
-const iotSiteWise = new IoTSiteWise(clientConfig);
 
 const DEFAULT_DASHBOARD_CONFIG = {
   displaySettings: {
@@ -42,7 +41,6 @@ const CLIENT_CONFIGURATION: DashboardClientConfiguration = {
   iotSiteWiseClient,
   iotEventsClient,
   iotTwinMakerClient,
-  iotSiteWise,
 };
 
 registerPlugin('metricsRecorder', {