+++ packages/react-components/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## [10.12.2](https://github.com/awslabs/iot-app-kit/compare/react-components-v10.12.1...react-components-v10.12.2) (2024-09-17)
+
+
+### Miscellaneous Chores
+
+* **react-components:** Synchronize iot-app-kit versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @iot-app-kit/components bumped from 10.12.1 to 10.12.2
+    * @iot-app-kit/core bumped from 10.12.1 to 10.12.2
+    * @iot-app-kit/core-util bumped from 10.12.1 to 10.12.2
+    * @iot-app-kit/source-iottwinmaker bumped from 10.12.1 to 10.12.2
+  * devDependencies
+    * @iot-app-kit/core bumped from 10.12.1 to 10.12.2
+    * @iot-app-kit/jest-config bumped from 10.12.1 to 10.12.2
+    * @iot-app-kit/source-iotsitewise bumped from 10.12.1 to 10.12.2
+    * @iot-app-kit/testing-util bumped from 10.12.1 to 10.12.2
+    * @iot-app-kit/ts-config bumped from 10.12.1 to 10.12.2
+    * eslint-config-iot-app-kit bumped from 10.12.1 to 10.12.2
+
 ## [10.12.1](https://github.com/awslabs/iot-app-kit/compare/react-components-v10.12.0...react-components-v10.12.1) (2024-09-14)